Make ErrorAlert a shared component
